refactor: CpuUsage 사용률 타입 변경
cpuPercentage 타입 Double -> BigDecimal으로 변경
소수점 자리처리와 더 나은 정밀도를 위한 타입수정

+++ task/src/main/java/com/terra/task/cpu/controller/CpuUsageController.java
@@ -2,6 +2,7 @@ package com.terra.task.cpu.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import java.math.BigDecimal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,7 +21,7 @@ public class CpuUsageController {
 
 	@Operation(summary = "CPU 사용량 조회", description = " cpu 사용량 조회 기능")
 	@GetMapping("/usage")
-	public double getCpuUsage() {
+	public BigDecimal getCpuUsage() {
 		return cpuUsageService.getCpuUsage();
 	}
